feat: added socket for zerion and isolated types
