+++ package.json
@@ -52,6 +52,7 @@
     "react-tooltip": "^4.2.21",
     "react-use-storage": "^0.4.4",
     "rebass": "^4.0.7",
+    "socket.io-client": "^2.4.0",
     "styled-components": "^5.3.3",
     "styled-icons": "^10.37.0",
     "typescript": "^4.3.5",
