+++ src/constants/types/contract.ts
@@ -1,0 +1,16 @@
+import { Contract } from '@ethersproject/contracts'
+
+export type ProductContract = {
+  name: string
+  contract: Contract
+}
+
+export type BondTellerContract = {
+  name: string
+  contract: Contract
+  isBondTellerErc20: boolean
+  isLp: boolean
+  underlyingAddr: string
+}
+
+export type ContractSources = { addr: string; abi: any }
