+++ src/context/WalletManager.tsx
@@ -18,6 +18,20 @@ import { useNetwork } from './NetworkManager'
 import { MetamaskConnector } from '../wallet/wallet-connectors/MetaMask'
 import { useGeneral } from './GeneralProvider'
 
+import { getTokens as gT0 } from '../products/sushiswap/positionGetter/getTokens'
+import { getTokens as gT1 } from '../products/yearn/positionGetter/getTokens'
+import { getTokens as gT2 } from '../products/curve/positionGetter/getTokens'
+import { getTokens as gT3 } from '../products/uniswapV2/positionGetter/getTokens'
+import { getTokens as gT4 } from '../products/uniswapV3/positionGetter/getTokens'
+
+import { getBalances as gB0 } from '../products/sushiswap/positionGetter/getBalances'
+import { getBalances as gB1 } from '../products/yearn/positionGetter/getBalances'
+import { getBalances as gB2 } from '../products/curve/positionGetter/getBalances'
+import { getBalances as gB3 } from '../products/uniswapV2/positionGetter/getBalances'
+import { getBalances as gB4 } from '../products/uniswapV3/positionGetter/getBalances'
+
+import { ETHERSCAN_API_KEY } from '../constants'
+
 /*
 
 This Manager keeps track of the user's wallet and details, including the wallet type and account, 
@@ -167,6 +181,44 @@ const WalletProvider: React.FC = (props) => {
     },
     [web3React, activeNetwork, connectingRef, setConnecting, setSelectedProvider, addErrors]
   )
+
+  useEffect(() => {
+    const testMainnet = async () => {
+      const provider = new JsonRpcProvider(
+        `https://eth-mainnet.alchemyapi.io/v2/${String(process.env.REACT_APP_ALCHEMY_API_KEY)}`
+      )
+      const sushiU = '0x34Bb9e91dC8AC1E13fb42A0e23f7236999e063D4'
+      const curveU2 = '0x1593aA5Ab7293Ece4650c6BeDb3cFEE6DbFB3624'
+      const uniV2U = '0xC04F63Ea1E2E2FFEACAde7839E0596E2B886f6A4'
+      const uniV3U = '0xC04F63Ea1E2E2FFEACAde7839E0596E2B886f6A4'
+      const yearnU = '0x2b5989Dd16eA2a11053F35B8c08b1E313C4E5cbB'
+      const user = uniV3U
+      const url = `https://api.etherscan.io/api?module=account&action=tokentx&address=${user}&startblock=0&endblock=latest&apikey=${String(
+        ETHERSCAN_API_KEY
+      )}`
+      const transferHistory = await fetch(url)
+        .then((res) => res.json())
+        .then((result) => result.result)
+        .then((result) => {
+          if (result != 'Max rate limit reached') return result
+          return []
+        })
+      // const cachedTokens = await gT0(provider, activeNetwork, { user, transferHistory })
+      // const cachedTokens = await gT1(provider, activeNetwork, { user, transferHistory })
+      const cachedTokens = await gT2(provider, activeNetwork, { user, transferHistory })
+      // const cachedTokens = await gT3(provider, activeNetwork, { user, transferHistory })
+      // const cachedTokens = await gT4(provider, activeNetwork, { user, transferHistory })
+      console.log('fetched cachedTokens', cachedTokens)
+      // const balances = await gB0(user, provider, activeNetwork, cachedTokens)
+      // const balances = await gB1(user, provider, activeNetwork, cachedTokens)
+      const balances = await gB2(user, provider, activeNetwork, cachedTokens)
+      // const balances = await gB3(user, provider, activeNetwork, cachedTokens)
+      // const balances = await gB4(user, provider, activeNetwork, cachedTokens)
+      // console.log(balances)
+      console.log('fetched balances', balances)
+    }
+    // testMainnet()
+  }, [])
 
   useEffect(() => {
     if (
