+++ src/products/sushiswap/positionGetter/getBalances.ts
@@ -16,7 +16,7 @@ import { ApolloClient, InMemoryCache, gql } from '@apollo/client'
 import { queryBalance } from '../../../utils/contract'
 import { getCoingeckoTokenPrice, getZapperProtocolBalances } from '../../../utils/api'
 import { WETH9_ADDRESS } from '../../../constants/mappings/tokenAddressMapping'
-import { createZapperBalanceMap, zapperNetworks } from '../../zapperBalances'
+import { createZapperBalanceMap, networkNames } from '../../zapperBalances'
 
 export const getBalances = async (
   user: string,
@@ -26,7 +26,7 @@ export const getBalances = async (
 ): Promise<Token[]> => {
   if (tokens.length == 0) return []
 
-  const zapperNet = zapperNetworks[activeNetwork.chainId]
+  const zapperNet = networkNames[activeNetwork.chainId]
   if (!zapperNet) return []
 
   const coinGeckoEthPrice = await getCoingeckoTokenPrice(WETH9_ADDRESS[activeNetwork.chainId], 'usd', 'ethereum')
