+++ src/products/uniswapV2/positionGetter/getBalances.ts
@@ -5,7 +5,7 @@ import { getContract } from '../../../utils'
 import { queryBalance } from '../../../utils/contract'
 import { ZERO } from '../../../constants'
 import { get1InchPrice, getCoingeckoTokenPrice, getZapperProtocolBalances } from '../../../utils/api'
-import { createZapperBalanceMap, zapperNetworks } from '../../zapperBalances'
+import { createZapperBalanceMap, networkNames } from '../../zapperBalances'
 import { WETH9_ADDRESS } from '../../../constants/mappings/tokenAddressMapping'
 
 export const getBalances = async (
@@ -45,7 +45,7 @@ export const getBalances = async (
   // }
   // return balances
 
-  const zapperNet = zapperNetworks[activeNetwork.chainId]
+  const zapperNet = networkNames[activeNetwork.chainId]
   if (!zapperNet) return []
 
   const coinGeckoEthPrice = await getCoingeckoTokenPrice(WETH9_ADDRESS[activeNetwork.chainId], 'usd', 'ethereum')
